pyproject.toml
[tool.poetry]
name = "school-feeding-project"
version = "0.1.0"
description = ""
authors = ["Arman <ghandilyanarman78@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.2"
djangorestframework = "^3.14.0"
pillow = "^10.2.0"
django-environ = "^0.11.2"
psycopg2 = "^2.9.9"
vt-py = "^0.18.0"
djoser = "^2.2.2"
pypdf2 = "^3.0.1"
python-docx = "^1.1.0"
python-pptx = "^0.6.22"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
